Make golangci-lint pass
A prior commit bootstrapped this repository with the boilerplate
golang-osd-operator convention, which uses golangci-lint in place of
gofmt and go vet. This commit addresses the errors called out by
golangci-lint. Some of the fixes are legit; some are spoofed, with TODOs
added, to be addressed later; and some (in tests) are simply bypassed.

Jira: [OSD-4943](https://issues.redhat.com/browse/OSD-4943)