Update dependencies
This patch updates the dependencies of the project. This is intended to
make it easier to import the API package from other projects. Currently
that is difficult because the code generated by the old version of the
`openapi-gen` tool isn't compatible with recent versions of the Kubernetes
API and the `controller-runtime` projects.

There are many changes in the `go.mod` and `go.sum` files, and in the
`vendor` directory. But the more relevant changes in the actual code of
the project are the following:

- Add the `+listType=atomic` annotation to the API types. In the new
  version of the `openapi-gen` tool this is mandatory for fields that
  are slices.

- Add the context parameter to the `Reconcile` method of the
  reconcilers. This is required in the new version of the
  `controller-runtime` project.

- Use `client.Object` instead of `runtime.Object`. This is required in
  the new version of the `controller-runtime` project.

- Use the `serviceusage` package instead of `servicemanagement` to get
  and enable Google API services. This is required in the new version of
  the Google API project.

- Adapte log, leadership and metrics configuration to the new version of
  the `controller-runtime` package.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>