Replace Go 1.16 with 1.17 in builder image
This patch changes the builder image so that it uses Go 1.17. Note that
there are currently no version of the _boilerplate_ base image that
contains Go 1.17, so the approach here is to remove Go 1.16 and install
1.17 explicitly.

Updating to Go 1.17 is necessary because without it it is not possible
to update to the latest versions of the Kubernetes API packages. In
particular the [sigs.k8s.io/json](https://github.com/kubernetes-sigs/json)
package doesn't work with Go 1.16 because it uses the `IsExported`
method that was added in Go 1.17.

Related: https://issues.redhat.com/browse/SDA-5500
Related: https://github.com/openshift/gcp-project-operator/pull/170
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>