Update dependencies
This patch updates the dependencies of the project. This is intended to
make it easier to import the API package from other projects. Currently
that is difficult because the code generated by the old version of the
`openapi-gen` tool isn't compatible with recent versions of the Kubernetes
API and the `controller-runtime` projects.

There are many changes in the `go.mod` and `go.sum` files, and in the
`vendor` directory. But the more relevant changes in the actual code of
the project are the following:

- Avoid importing the `operator-sdk` project. That isn't necessary with
  the new version of the Operator SDK. Applications that need components
  from the SDK can import the `operator-lib` project, but this application
  doesn't even need that.

- Add the `+listType=atomic` annotation to the API types. In the new
  version of the `openapi-gen` tool this is mandatory for fields that
  are slices.

- Add the context parameter to the `Reconcile` method of the
  reconcilers. This is required in the new version of the
  `controller-runtime` project.

- Use `client.Object` instead of `runtime.Object`. This is required in
  the new version of the `controller-runtime` project.

- Use the `serviceusage` package instead of `servicemanagement` to get
  and enable Google API services. This is required in the new version of
  the Google API project.

- Adapt log, leadership and metrics configuration to the new version of
  the `controller-runtime` package.

Unfortunately these changes can't be done in separate patches because
updating a single dependency pulls many other dependencies, and
introduces conflicts.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>